Implemented Repository Pattern

+++ NicasourceChallenge.Core/Entities/Document.cs
@@ -9,7 +9,7 @@ public class Document : BaseEntity
     public string? Name { get; set; }
     public string? Description { get; set; }
     public string? Url { get; set; }
-    public float Size { get; set; }
+    public long Size { get; set; }
     public string? Format { get; set; }
 }
